M4-sponsor-gcp/requirements.txt
functions-framework==3.4.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
google-cloud-storage==2.14.0
